:adhesive_bandage: (sonarcloud): Fix implicit conversion from 'unsigned int' to 'float' may lose precision

+++ spikes/lk_microphone/main.cpp
@@ -64,7 +64,7 @@ int RMS(std::array<int, 10> &data, int newvalue)
 
 	square += newvalue * newvalue;
 
-	mean = square / data.size();
+	mean = static_cast<float>(square) / static_cast<float>(data.size());
 
 	root = std::sqrt(mean);
